add rankstr library, more MPI examples
