+++ src/CMake/Setup3rdParty.cmake
@@ -6,6 +6,10 @@
 include(CMake/thirdparty/SetupHwloc.cmake)
 include(CMake/thirdparty/SetupJansson.cmake)
 
+if(ENABLE_MPI)
+    include(CMake/thirdparty/SetupRankstr.cmake)
+endif()
+
 if(BUILD_DOCS)
     find_package(Doxygen)
     include(CMake/thirdparty/FindSphinx.cmake)
