+++ src/examples/mpi-openmp-examples/CMakeLists.txt
@@ -17,8 +17,8 @@ message(STATUS "Adding variorum MPI examples")
 foreach(EXAMPLE ${MPI_EXAMPLES})
     message(STATUS " [*] Adding MPI example: ${EXAMPLE}")
     add_executable(${EXAMPLE} ${EXAMPLE}.c)
-    target_include_directories(${EXAMPLE} PUBLIC ${MPI_CXX_INCLUDE_PATH})
-    target_link_libraries(${EXAMPLE} variorum ${variorum_deps} ${MPI_CXX_LIBRARIES} ${MPI_CXX_LINK_FLAGS})
+    target_include_directories(${EXAMPLE} PUBLIC ${MPI_CXX_INCLUDE_PATH} ${RANKSTR_INCLUDE_DIRS})
+    target_link_libraries(${EXAMPLE} variorum ${variorum_deps} ${MPI_CXX_LIBRARIES} ${MPI_CXX_LINK_FLAGS} ${RANKSTR_LIBRARY})
 endforeach()
 
 include_directories(${CMAKE_SOURCE_DIR}/variorum)
