+++ src/examples/mpi-openmp-examples/runme.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-srun -N 2 ./variorum-print-power-mpi-example
+# Launch 8 tasks, 4 per node using Slurm. The result is 1 printout per node of
+# the power usage.
+srun -N 2 -n 8 ./variorum-print-power-mpi-example
 
 #
