+++ src/examples/mpi-openmp-examples/variorum-print-power-limits-mpi-example.c
@@ -5,21 +5,32 @@
 
 #include <stdio.h>
 #include <mpi.h>
+#include <rankstr_mpi.h>
+#include <unistd.h>
 
 #include <variorum.h>
 
 int main(int argc, char **argv)
 {
     int ret;
-    int numprocs, rank;
+    int numprocs, rank, len;
+    char host[1024];
 
     MPI_Init(NULL, NULL);
     MPI_Comm_size(MPI_COMM_WORLD, &numprocs);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+    gethostname(host, 1023);
+
+    MPI_Comm newcomm;
+    int new_rank, new_size;
+
+    rankstr_mpi_comm_split(MPI_COMM_WORLD, host, rank, 1, 2, &newcomm);
+    MPI_Comm_size(newcomm, &new_size);
+    MPI_Comm_rank(newcomm, &new_rank);
 
     // higher-level software must check for thread and process safety
-    // we assume rank 0 is responsible for monitor and control
-    if (rank == 0)
+    // we assume rank 0 on each node is responsible for monitor and control
+    if (new_rank == 0)
     {
         ret = variorum_print_power_limits();
         if (ret != 0)
@@ -28,6 +39,8 @@ int main(int argc, char **argv)
         }
     }
 
+    MPI_Bcast(&ret, 1, MPI_INT, new_rank, MPI_COMM_WORLD);
+    MPI_Comm_free(&newcomm);
     MPI_Finalize();
     return ret;
 }
